+++ src/grbl/spindle_control.h
@@ -49,10 +49,10 @@ uint8_t spindle_get_state();
   
   // Sets spindle PWM quickly for stepper ISR. Also called by spindle_set_state().
   // NOTE: 328p PWM register is 8-bit.
-  void spindle_set_speed(uint8_t pwm_value);
+  void spindle_set_speed(uint16_t pwm_value);
   
   // Computes 328p-specific PWM register value for the given RPM for quick updating.
-  uint8_t spindle_compute_pwm_value(float rpm);
+  uint16_t spindle_compute_pwm_value(float rpm);
   
 #else
